Text.ParserCombinators.Poly.Parser

Description

This module contains the definitions for a generic parser, without running state. These are the parts that are shared between the Plain and Lazy variations. Do not import this module directly, but only via T.P.Poly.Plain or T.P.Poly.Lazy.

The Parser datatype

newtype Parser t a

This Parser datatype is a fairly generic parsing monad with error reporting. It can be used for arbitrary token types, not just String input. (If you require a running state, use module Poly.State instead)

Constructors

P ([t] -> Result [t] a)

data Result z a

A return type like Either, that distinguishes not only between right and wrong answers, but also has commitment, so that a failure cannot be undone. This should only be used for writing very primitive parsers - really it is an internal detail of the library. The z type is the remaining unconsumed input.

Constructors

Success z a

Failure z String

Committed (Result z a)

Basic parsers

next :: Parser t t

Simply return the next token in the input tokenstream.

eof :: Parser t ()

Succeed if the end of file/input has been reached, fail otherwise.

satisfy :: (t -> Bool) -> Parser t t

Return the next token if it satisfies the given predicate.

satisfyMsg :: Show t => (t -> Bool) -> String -> Parser t t

Return the next token if it satisfies the given predicate. The String argument describes the function, for better error messages.

onFail :: Parser t a -> Parser t a -> Parser t a infixl 6

p onFail q means parse p, unless p fails, in which case parse q instead. Can be chained together to give multiple attempts to parse something. (Note that q could itself be a failing parser, e.g. to change the error message from that defined in p to something different.) However, a severe failure in p cannot be ignored.

Re-parsing

reparse :: [t] -> Parser t ()

Push some tokens back onto the front of the input stream and reparse. This is useful e.g. for recursively expanding macros. When the user-parser recognises a macro use, it can lookup the macro expansion from the parse state, lex it, and then stuff the lexed expansion back down into the parser.