Text.ParserCombinators.Poly.Result

The parsing result type

data Result z a

A return type like Either, that distinguishes not only between right and wrong answers, but also has commitment, so that a failure cannot be undone. This should only be used for writing very primitive parsers - really it is an internal detail of the library. The z type is the remaining unconsumed input.

Constructors

Success z a

Failure z String

Committed (Result z a)

resultToEither :: Result z a -> (Either String a, z)

Convert a Result to an Either, paired with the remaining unconsumed input.