Text.ParserCombinators.Poly.StateLazy

The Parser datatype

newtype Parser s t a

The only differences between a State and a StateLazy parser are the instance of Applicative, and the type (and implementation) of runParser. We therefore need to newtype the original Parser type, to allow it to have a different instance.

Constructors

P (Parser s t a)

data Result z a

A return type like Either, that distinguishes not only between right and wrong answers, but also has commitment, so that a failure cannot be undone. This should only be used for writing very primitive parsers - really it is an internal detail of the library. The z type is the remaining unconsumed input.

Constructors

Success z a

Failure z String

Committed (Result z a)

runParser :: Parser s t a -> s -> [t] -> (a, s, [t])

Apply a parser to an input token sequence.

Basic parsers

next :: Parser s t t

Simply return the next token in the input tokenstream.

eof :: Parser s t ()

Succeed if the end of file/input has been reached, fail otherwise.

satisfy :: (t -> Bool) -> Parser s t t

Return the next token if it satisfies the given predicate.

onFail :: Parser s t a -> Parser s t a -> Parser s t a

p onFail q means parse p, unless p fails, in which case parse q instead. Can be chained together to give multiple attempts to parse something. (Note that q could itself be a failing parser, e.g. to change the error message from that defined in p to something different.) However, a severe failure in p cannot be ignored.

manyFinally :: Parser s t a -> Parser s t z -> Parser s t [a]

State-handling

stUpdate :: (s -> s) -> Parser s t ()

Update the internal state.

stQuery :: (s -> a) -> Parser s t a

Query the internal state.

stGet :: Parser s t s

Deliver the entire internal state.

Re-parsing

reparse :: [t] -> Parser s t ()

Push some tokens back onto the front of the input stream and reparse. This is useful e.g. for recursively expanding macros. When the user-parser recognises a macro use, it can lookup the macro expansion from the parse state, lex it, and then stuff the lexed expansion back down into the parser.

Re-export all more general combinators

module Text.ParserCombinators.Poly.Base

module Control.Applicative