Text.ParserCombinators.Poly.StateText

The Parser datatype

newtype Parser s a

This Parser datatype is a specialised parsing monad with error reporting. Whereas the standard version can be used for arbitrary token types, this version is specialised to Text input only.

Constructors

P (s -> Text -> Result (Text, s) a)

data Result z a

A return type like Either, that distinguishes not only between right and wrong answers, but also has commitment, so that a failure cannot be undone. This should only be used for writing very primitive parsers - really it is an internal detail of the library. The z type is the remaining unconsumed input.

Constructors

Success z a

Failure z String

Committed (Result z a)

runParser :: Parser s a -> s -> Text -> (Either String a, s, Text)

Apply a parser to an input token sequence.

Basic parsers

next :: Parser s Char

Simply return the next token in the input tokenstream.

eof :: Parser s ()

Succeed if the end of file/input has been reached, fail otherwise.

satisfy :: (Char -> Bool) -> Parser s Char

Return the next token if it satisfies the given predicate.

onFail :: Parser s a -> Parser s a -> Parser s a

p onFail q means parse p, unless p fails, in which case parse q instead. Can be chained together to give multiple attempts to parse something. (Note that q could itself be a failing parser, e.g. to change the error message from that defined in p to something different.) However, a severe failure in p cannot be ignored.

Derived parsers (but implemented more efficiently)

manySatisfy :: (Char -> Bool) -> Parser s Text

manySatisfy p is a more efficient fused version of many (satisfy p)

many1Satisfy :: (Char -> Bool) -> Parser s Text

many1Satisfy p is a more efficient fused version of many1 (satisfy p)

State-handling

stUpdate :: (s -> s) -> Parser s ()

Update the internal state.

stQuery :: (s -> a) -> Parser s a

Query the internal state.

stGet :: Parser s s

Deliver the entire internal state.

Re-parsing

reparse :: Text -> Parser s ()

Push some tokens back onto the front of the input stream and reparse. This is useful e.g. for recursively expanding macros. When the user-parser recognises a macro use, it can lookup the macro expansion from the parse state, lex it, and then stuff the lexed expansion back down into the parser.

Re-export all more general combinators

module Text.ParserCombinators.Poly.Base

module Control.Applicative